+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.3.0 (WIP)
 
-* Converted from an ad-hoc lifecycle management solution to integrant.
+* Converted whole app from an ad-hoc lifecycle management solution to integrant.
 * Added a mandatory `:id` property to sites.
+* Introduced an explicit task scheduler for checking sites, allowing for concurrent
+  checks and easy on-demand website checks.
+* Added a mandatory `[:state :loading?]` property to sites.